refactor(processing): move StreamThrottlePolicy to model package
Separate the policy data model from the interface definition by
relocating StreamThrottlePolicy from api.processing to api.model.processing,
consistent with other model types (retry, config).
