+++ stream-android-core/src/test/java/io/getstream/android/core/internal/processing/StreamThrottlerImplTest.kt
@@ -18,7 +18,7 @@
 
 package io.getstream.android.core.internal.processing
 
-import io.getstream.android.core.api.processing.StreamThrottlePolicy
+import io.getstream.android.core.api.model.processing.StreamThrottlePolicy
 import io.mockk.mockk
 import kotlin.test.assertFailsWith
 import kotlinx.coroutines.CoroutineExceptionHandler
